adding new class for handling the service with the controller of OCR, adding input and output names
